refactor(middleware): rename *_pi to canonical names + drop Pi class suffix (M6.1)
- Rename 11 middleware modules (artifacts/attachments/citation/memory/sandbox/
  skills/subagents/timestamps/todo/cost/compaction) from *_pi.py to canonical.
- Merge standalone compaction_pi.py into compaction/__init__.py alongside the
  existing boundary/summarizer/tokens helpers.
- Drop `Pi` suffix from all middleware classes (ArtifactMiddlewarePi →
  ArtifactMiddleware, etc.).
- Update imports in run_manager.py, citations/__init__.py, and all unit tests.
- Rename test files tests/unit/test_*_pi.py → tests/unit/test_*.py.
