+++ backend/tests/unit/test_cost.py
@@ -1,4 +1,4 @@
-"""Unit tests for CostMiddlewarePi (M3.d.1).
+"""Unit tests for CostMiddleware (M3.d.1).
 
 Covers:
 - Constructor stores all fields correctly.
@@ -19,7 +19,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 from cubepi.providers.base import AssistantMessage, TextContent, Usage
 
-from cubebox.middleware.cost_pi import CostMiddlewarePi, _extract_usage
+from cubebox.middleware.cost import CostMiddleware, _extract_usage
 
 # ---------------------------------------------------------------------------
 # Helpers
@@ -62,7 +62,7 @@ def _make_response_no_usage(
     )
 
 
-def _make_middleware(**kwargs: Any) -> CostMiddlewarePi:
+def _make_middleware(**kwargs: Any) -> CostMiddleware:
     defaults: dict[str, Any] = {
         "org_id": "org-1",
         "workspace_id": "ws-1",
@@ -70,7 +70,7 @@ def _make_middleware(**kwargs: Any) -> CostMiddlewarePi:
         "conversation_id": "conv-1",
     }
     defaults.update(kwargs)
-    return CostMiddlewarePi(**defaults)
+    return CostMiddleware(**defaults)
 
 
 # ---------------------------------------------------------------------------
@@ -79,7 +79,7 @@ def _make_middleware(**kwargs: Any) -> CostMiddlewarePi:
 
 
 def test_constructor_stores_all_fields() -> None:
-    mw = CostMiddlewarePi(
+    mw = CostMiddleware(
         org_id="org-abc",
         workspace_id="ws-xyz",
         user_id="usr-42",
@@ -119,7 +119,7 @@ async def test_after_model_response_returns_none() -> None:
     response = _make_response()
     ctx = MagicMock()
 
-    with patch("cubebox.middleware.cost_pi.asyncio.create_task"):
+    with patch("cubebox.middleware.cost.asyncio.create_task"):
         result = await mw.after_model_response(response, ctx)
 
     assert result is None
@@ -139,7 +139,7 @@ async def test_after_model_response_updates_last_billing_id() -> None:
     response = _make_response()
     ctx = MagicMock()
 
-    with patch("cubebox.middleware.cost_pi.asyncio.create_task"):
+    with patch("cubebox.middleware.cost.asyncio.create_task"):
         await mw.after_model_response(response, ctx)
 
     assert mw._last_billing_id is not None
@@ -153,7 +153,7 @@ async def test_after_model_response_advances_billing_id_each_call() -> None:
     ctx = MagicMock()
     response = _make_response()
 
-    with patch("cubebox.middleware.cost_pi.asyncio.create_task"):
+    with patch("cubebox.middleware.cost.asyncio.create_task"):
         await mw.after_model_response(response, ctx)
         first_id = mw._last_billing_id
 
@@ -171,7 +171,7 @@ async def test_after_model_response_advances_billing_id_each_call() -> None:
 @pytest.mark.asyncio
 async def test_after_model_response_writes_billing_record_with_correct_attribution() -> None:
     """_write is called with correct org/workspace/user/conversation attribution."""
-    mw = CostMiddlewarePi(
+    mw = CostMiddleware(
         org_id="org-billing",
         workspace_id="ws-billing",
         user_id="usr-billing",
@@ -205,8 +205,8 @@ async def test_after_model_response_writes_billing_record_with_correct_attributi
     fake_session.__aexit__ = AsyncMock(return_value=False)
 
     with (
-        patch("cubebox.middleware.cost_pi.async_session_maker", return_value=fake_session),
-        patch("cubebox.middleware.cost_pi.BillingRepository", _FakeRepo),
+        patch("cubebox.middleware.cost.async_session_maker", return_value=fake_session),
+        patch("cubebox.middleware.cost.BillingRepository", _FakeRepo),
     ):
         await mw.after_model_response(response, ctx)
         # Allow the fire-and-forget task to complete
@@ -260,8 +260,8 @@ async def test_after_model_response_billing_event_id_matches_llm_event() -> None
     fake_session.__aexit__ = AsyncMock(return_value=False)
 
     with (
-        patch("cubebox.middleware.cost_pi.async_session_maker", return_value=fake_session),
-        patch("cubebox.middleware.cost_pi.BillingRepository", _FakeRepo),
+        patch("cubebox.middleware.cost.async_session_maker", return_value=fake_session),
+        patch("cubebox.middleware.cost.BillingRepository", _FakeRepo),
     ):
         await mw.after_model_response(response, ctx)
         await asyncio.sleep(0)
@@ -290,8 +290,8 @@ async def test_after_model_response_id_matches_last_billing_id() -> None:
     fake_session.__aexit__ = AsyncMock(return_value=False)
 
     with (
-        patch("cubebox.middleware.cost_pi.async_session_maker", return_value=fake_session),
-        patch("cubebox.middleware.cost_pi.BillingRepository", _FakeRepo),
+        patch("cubebox.middleware.cost.async_session_maker", return_value=fake_session),
+        patch("cubebox.middleware.cost.BillingRepository", _FakeRepo),
     ):
         await mw.after_model_response(response, ctx)
         await asyncio.sleep(0)
@@ -367,8 +367,8 @@ async def test_billing_write_failure_does_not_raise() -> None:
     fake_session.__aexit__ = AsyncMock(return_value=False)
 
     with (
-        patch("cubebox.middleware.cost_pi.async_session_maker", return_value=fake_session),
-        patch("cubebox.middleware.cost_pi.BillingRepository", _BrokenRepo),
+        patch("cubebox.middleware.cost.async_session_maker", return_value=fake_session),
+        patch("cubebox.middleware.cost.BillingRepository", _BrokenRepo),
     ):
         result = await mw.after_model_response(response, ctx)
         await asyncio.sleep(0)
